backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
asyncpg==0.29.0
pydantic-settings==2.3.4
sqlalchemy>=2.0.0
alembic
vectorbt
quantstats
PyPortfolioOpt
empyrical-reloaded
pandas-ta
yfinance
openbb-core
alpaca-py
anthropic
redis
celery
